test: split finalizer and product suites
